Extras/requirements.txt
pytest==8.3.3
pytest_bdd==8.1.0
python-dotenv==1.0.1
selenium==4.27.1
